Add capybara support, with akephalos headless browser.
There seems to be some weirdness with the service as it doesn't behave
well at all with nil username/password values.  It's as if EM gets
hung and never shuts down.  It also takes a surprisingly long time for
the service to undeploy.  Not sure where the bottleneck is, though.

Because the app doesn't *really* depend on the service, you can omit
service from the deploy command and things will run smoother and
faster, nil un/pw or not.

Still need to find out why Rails apps take so long to deploy.